perf: Skip calculating individual account balances
Using the standard barchart report as a starting point, this report
previously used the standard report's individual account balances to
generate a list of grouped balances per tag value. This was obviously
inefficient, since per-account granularity is not needed for this report.

Now, grouped balances are calculated on the primary account traversal.
On my system, the report renders about 25% faster with this change.